no trailing space when joining prefix/suffix cmds
Join prefix and suffix commands using "&& ", do not use a trailing space.
In some scriping languages, such as batch, a trailing space can change the
behavior of a command.  The particular cases where this was observed was in the
command "set TEMP=%CD%".  When joined with a trailing space the space was added
to the TEMP variable.

See: https://superuser.com/questions/57723