Fix rsync-build-pack: 'resource is not used' error
Concourse doesn't like when a resource is defined and not used. The fly tool
exits straight away. A very simple way to trigger this is to spawn a pipeline
with the argument: '-p linux-*'. This patch fixes this bug by restricting the
resource definition to the case when the pipeline is supposed to run on atleast
one of osx, win.