optional use of a pool of locks to limit builds
Optionally use a pool of locks to limit the number of parallel builds.
A lock for a particular platform is aquired prior to performing the
build task.  This lock is released regardless of the success of the
build.

Relevant documentation:
pool-resource: https://github.com/concourse/pool-resource
ensure: https://concourse-ci.org/ensure-step-hook.html