lock pool repo can be auth via user/pass or key
Access to the lock pool repository can be done via a private key, a
username and password, or using no authentication.  Private key auth is
preferred and will be used when the configuration file contain a
lock-pool-private-key entry.  Username and password is used when a no
private key is specified but a lock-pool-username entry is present.