add --use-repo-access arg to one-off and batch
Add the --use-repo-access argument to the one-off and batch subcommands.
This argument conditionally enables the repo access credentials
(username and token) to be passed to the build jobs. By default these
are not passed.

Writing the _netrc/.netrc can stall workers if another job is accessing or
writing the file at the same time. This effects windows especially. The
argument will not write these files unless they are specifically
required and requested by the user.