add optional stage_for_upload job
Add an optional "stage_for_upload" jobs to pipelines which executes the
task defined in the stage_for_upload_config section of the configuration
with the artifact directory passes in via the ARTIFACT_DIR environment
variable. This job is created when the --stage-for-upload option is
specified on the command line.

Additionally, adds a --commit-msg command line option where a git commit
message can be specified that is passed to the the upload job via the
GIT_COMMIT_MSG environment variable.