cli option to add push-branch job
Add a command line option to add a job to the concourse pipeline
responsible for pushing a named branch to master. The actual logic for
the push is defined in the push-branch-config section of the
configuration file.

Example usage:
c3i one-off prefix_sphinx --automated-pipeline --branches test --push-branch --stage-for-upload --commit-msg "sphinx 3.0.4" sphinx-feedstock